Version 0.0.5
Added the rest of the keyboard characters to the alphabet array. as well
as fixed an issue where a character in the key would be skipped every
time a space or miscellaneous character was detected in the text array.

known issues:
-additional character in the array encrypted as well as encrypted
missing a character from the array text
-un-necessary additional spaces when outputing either the Key or the
Encrypted text, as well as 2 extra spaces for every space in the
original text